+++ src/store/index.ts
@@ -1,3 +1,6 @@
 export { store } from './root';
 export { persistor } from './root';
 export type { RootState } from './root';
+
+export { fetchUser } from './user';
+export { fetchRoutine } from './routine';
